+++ src/jiratui/config.py
@@ -94,6 +94,9 @@ class ApplicationConfiguration(BaseSettings):
     """Set this to False if your Jira instance run on-premises."""
     use_bearer_authentication: bool = False
     """Set this to `True` if your Jira instance uses Bearer authentication instead of Basic authentication."""
+    use_cert_authentication: bool = False
+    """Set this to `True` if your Jira instance uses certificate-based authentication instead of Bearer authentication
+    or Basic authentication."""
     jira_user_group_id: str | None = None
     """**[DEPRECATED]** The ID of the group that contains all (or most) of the Jira users in your Jira installation. This
     value is used as a fall back mechanism to fetch available users. This is only supported in the Jira Cloud
@@ -205,7 +208,7 @@ class ApplicationConfiguration(BaseSettings):
     """When this is `True` JiraTUI will use Jira ability to do full-text search not only in summary and description
     fields but in any text-based field, including comments. This may be slower. If this is False JiraTUI will only
     search items by summary and description fields."""
-    ssl: SSLConfiguration | None = Field(default_factory=SSLConfiguration)
+    ssl: SSLConfiguration | None = Field(default=None)
     """SSL configuration for client-side certificates and CA bundle."""
     search_results_default_order: WorkItemsSearchOrderBy = WorkItemsSearchOrderBy.CREATED_DESC
     """The default order for search results. Accepts values from [WorkItemsSearchOrderBy](#jiratui.models.WorkItemsSearchOrderBy) enum: `CREATED_ASC`,