+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Add a read-only mode that prevents JiraTUI from sending mutation requests. By [@tagplus5](https://github.com/tagplus5) in https://github.com/whyisdifficult/jiratui/pull/
 - Add a new CLI command `jiratui issues clone` to clone a work item. By [@whyisdifficult](https://github.com/whyisdifficult) in https://github.com/whyisdifficult/jiratui/pull/
 - Add a new CLI command `jiratui issues delete` to delete a work item. By [@whyisdifficult](https://github.com/whyisdifficult) in https://github.com/whyisdifficult/jiratui/pull/