+++ CHANGELOG.md
@@ -10,8 +10,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add a read-only mode that prevents JiraTUI from sending mutation requests. By [@tagplus5](https://github.com/tagplus5) in https://github.com/whyisdifficult/jiratui/pull/
+
+## [1.12.0] 2026-08-01
+
+### Added
+
 - Add a new CLI command `jiratui issues clone` to clone a work item. By [@whyisdifficult](https://github.com/whyisdifficult) in https://github.com/whyisdifficult/jiratui/pull/
 - Add a new CLI command `jiratui issues delete` to delete a work item. By [@whyisdifficult](https://github.com/whyisdifficult) in https://github.com/whyisdifficult/jiratui/pull/
+- Add new configuration option `use_cert_authentication`, to load client certs into SSLContext. By [wbh1](https://github.com/wbh1) in https://github.com/whyisdifficult/jiratui/pull/157
 
 ## [1.11.2] 2026-07-25